+++ AlpacaService/Domain/Services/Implementation/SendQuoteJob.cs
@@ -1,6 +1,3 @@
-using BN.PROJECT.Core;
-using Microsoft.CodeAnalysis.Elfie.Model;
-
 namespace BN.PROJECT.AlpacaService;
 
 [PersistJobDataAfterExecution]
@@ -10,24 +7,29 @@ public class SendQuoteJob : IJob
     private readonly ILogger<SendQuoteJob> _logger;
     private readonly IConfiguration _configuration;
     private readonly IAlpacaDataService _alpacaDataService;
-    private readonly IAlpacaTradingService _alpacaTradingService;
     private readonly IAlpacaRepository _alpacaRepository;
     private readonly IKafkaProducerService _kafkaProducer;
+    private readonly IHubContext<AlpacaHub> _hubContext;
+    private readonly IDatabase _redisDatabase;
+
 
     public SendQuoteJob(
         ILogger<SendQuoteJob> logger,
         IConfiguration configuration,
         IAlpacaDataService alpacaDataService,
-        IAlpacaTradingService alpacaTradingService,
         IAlpacaRepository alpacaRepository,
-        IKafkaProducerService kafkaProducer)
+        IKafkaProducerService kafkaProducer,
+        IHubContext<AlpacaHub> hubContext,
+        IConnectionMultiplexer redis)
     {
         _logger = logger;
         _configuration = configuration;
         _alpacaDataService = alpacaDataService;
         _alpacaRepository = alpacaRepository;
-        _alpacaTradingService = alpacaTradingService;
         _kafkaProducer = kafkaProducer;
+        _hubContext = hubContext;
+        _redisDatabase = redis.GetDatabase();
+
     }
 
     public async Task Execute(IJobExecutionContext context)
@@ -69,6 +71,21 @@ public class SendQuoteJob : IJob
             };
 
             await _kafkaProducer.SendMessageAsync("strategy", message.ToJson());
+
+            var quoteMessage = new QuoteMessage
+            {
+                UserId = exec.UserId,
+                Symbol = quote.Symbol,
+                AskPrice = quote.AskPrice,
+                BidPrice = quote.BidPrice,
+                TimestampUtc = quote.TimestampUtc
+            };
+
+            var connectionId = await _redisDatabase.StringGetAsync(exec.UserId.ToString());
+            if (!string.IsNullOrEmpty(connectionId.ToString()))
+            {
+                await _hubContext.Clients.Client(connectionId).SendAsync("ReceiveQuote", message);
+            }
         }
 
         _logger.LogInformation("Instance " + key + " Send Quote Job end");