+++ AlpacaService/GlobalUsings.cs
@@ -15,3 +15,5 @@ global using System.ComponentModel.DataAnnotations.Schema;
 global using System.Net.Http.Headers;
 global using System.Text;
 global using System.Collections.Concurrent;
+global using Microsoft.AspNetCore.SignalR;
+global using StackExchange.Redis;