+++ AlpacaService/Release-Notes.md
@@ -2,6 +2,10 @@
 
 BN.PROJECT is a platform which combines brokerage operations with algorithmic trading.
 
+## Version 2025.04.1
+
+- Redis cache added
+- Ratefeed via SignalR 
 
 ## Version 2025.03.3